+++ CHANGELOG.md
@@ -177,4 +177,5 @@
 * Added RpgInventoryModified messages sent when inventory is changed
 * Fixed OnFire and OnReload scripts being passed item model instead of item instance
 * Melee weapon damage can be delayed no longer WIP
+* Moved input lock and player in control checks in weapon controller to a more correct spot
 * Added recock handling to weapons [WIP]